[Swift+WASM] [IRGen] Disable COMDAT for reflection metadata
LLVM's Wasm COMDAT code assumes each COMDAT will be in its own section.
This works for C++ on Clang, since each COMDAT is emitted into its own
.rodata.<symbol name> section,
but not for Swift metadata, which are all placed in the same section.

Disable COMDAT support when emitting reflection metadata.

This matches the behaviour on ELF (where COMDATs are disabled) and MachO
(which doesn't support COMDAT)

This shouldn't affect existing platforms.