[CMake] Replace swift component installation with CMake built-ins
The current state of things is that install rules for a target, file,
or directory are only created if the component they are associated with
is being installed. However, CMake has the notion of components
associated with installation.

This commit's intent is not to get rid of the swift component system,
but rather remove some custom support code built in favor of using
CMake's built-in support.