Delete WindowsSubsystemForLinux.md
This is duplicated by the Linux rules.  The Windows Subsystem for Linux is not a unique environment, it is an Ubuntu environment running with a bridge into the NT kernel.  There is nothing to be done to support it, just follow the Linux rules.  The Windows documentation has been updated to point users to the Linux section.