Change from _ContiguousCollection to AccelerateBuffer.
Since we are not adding ContiguousCollection to the standard library at this point in time, we can and should use something more narrowly tailored to Accelerate's needs. In particular, Accelerate does not require almost anything from Collection; it only needs withUnsafe[Mutable]BufferPointer, count, and the Element associatedtype. So let's pare the protocol back to only those things, and give it a name scoped as well: [Mutable]AccelerateBuffer.

In practice, most types people use with these API will be Collections anyway, but this enables a few useful additional use cases:

- Fixed-size buffers
- Retroactive conformance of SIMD types