AST: inline the definition for `AnyValue::Holder<Type>::equals`
Unfortunately, MSVC will not preserve the template specialization in
swiftAST.  However, the function is pretty tiny, so simply inline it
into the header.  This allows the template specialization to be
instantiated and inlined at the use site.  The overall size impact
should be negligible and this repairs the Windows build.