[TypeChecker] [NFC] Make unreachable case unreachable
A part of swift::diagnoseExplicitUnavailability() that should never run for deprecations looks for them and handles them with “break”. Change this to llvm_unreachable() so we notice if this somehow happens.

This change *should* be NFC; I’m splitting it out from other changes so that bisections can point to it if it causes problems in the future.