[sil] Teach the verifier that open_existential_addr immutable is safe to pass as Indirect_InoutAliasable.
We do not consider inout_aliasable to be "truly mutating" since today it is just
used as a way to mark a captured argument and not that something truly has
mutating semantics. The reason why this is safe is that the typechecker
guarantees that if our value was immutable, then the use in the closure must be
immutable as well.

In a future SIL, we want to remove Inout_Aliasable in favor of just using
inout/in_guaranteed using the capture info from the type checker.

rdar://50212579
(cherry picked from commit 437475421798bca3374acdeebdd348d537254933)