SIL: Set associated DC of a SILModule correctly in SIL mode
When the frontend is run with a .sil input we take a different code
path that created a SILModule whose associated context was the
ModuleDecl, even for a non-WMO build. Make sure we pass in a
SourceFile in this case.

We don't allow multiple input files when we have a .sil input
at all, so WMO does not change anything in how the driver calls
teh frontend.

However, it does change optimizer behavior, so we can't just
unconditionally set WholeModule to true for a frontend
invocation with a .sil input.