Un-XFAIL validation-test/Runtime/class_stubs.m and adjust for Clang changes
The objc_loadClassref() symbol is weak linked from Clang now, so we can
un-XFAIL this test. Note that we have to pass a special flag to Clang
since objc_loadClassref() is not in the SDK's TBD files yet.

Also, change the test to use its own asserts instead of FileCheck since
there's no output unless a new libobjc is available.