[android] Modify lit replacement of module-target-triple for ARMv7
Android ARMv7 has an LLVM triple of armv7-none-linux-androideabi, but
the Swift module triple is just armv7-node-linux-android. The change
removes the "eabi" bit for the %module-target-triple substitution of lit
to avoid an error in the test
test/SourceKit/CursorInfo/cursor_swiftonly_systemmodule.swift.

Android AArch64 wasn't affected because the LLVM triple for it is just
aarch64-none-linux-android.