[silgenpattern] When binding values, use a ManagedValue instead of an RValue to avoid implicitly exploding tuples.
In SILGenPattern, we need to be able to unforward cleanups when we explode
tuples. Thus we can't use RValue in SILGenPattern since it may implicitly
explode tuples (which without modifying RValue itself we can not
unforward). This patch removes a specific RValue usage that we can replace with
the use of a ManagedValue instead.

rdar://49903264
(cherry picked from commit adcfda6f6db4594c6bf8fef116fdb1ed00812f5a)

Conflicts:
	test/Inputs/resilient_struct.swift

The conflict was that on master, we added a few more types to the input file. I
just added them.